CA Fixes: CA1401 - accessibility of P/Invoke
* CA2214 - GlobalKeys() -> CreateHandle()
* CA1063 - the disposable pattern
* CA2213 - Designer Dispose
* CA2111 - Pointers should not be visible
* CA1901 - P/Invoke declarations

+++ TmVTweaks/WndEx.cs
@@ -23,7 +23,7 @@ namespace net.r_eg.TmVTweaks
 
         protected void setWindowStyle(IntPtr hWnd, long style)
         {
-            NativeMethods.SetWindowLong(hWnd, GWL.GWL_STYLE, style);
+            NativeMethods.SetWindowLong(hWnd, GWL.GWL_STYLE, (int)style);
         }
 
         protected bool isVisible(IntPtr hWnd)
